requirements.txt
aiosmtplib==2.0.1
aiosqlite==0.17.0
anyio==3.7.0
asyncpg==0.27.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-mail==1.2.8
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
passlib==1.7.4
pydantic==1.10.8
PyJWT==2.7.0
pypika-tortoise==0.1.6
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.19.3
typing-extensions==4.6.2
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
